// opencdx-dashboard-new/src/api/communications/model/notification-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { Timestamp } from './timestamp';

/**
 * 
 * @export
 * @interface NotificationEvent
 */
export interface NotificationEvent {
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'eventId'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'eventName'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'eventDescription'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'emailTemplateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'smsTemplateId'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'messageTemplateId'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof NotificationEvent
     */
    'eventParameters'?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'priority'?: Priority;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'sensitivity'?: Sensitivity;
    /**
     * 
     * @type {number}
     * @memberof NotificationEvent
     */
    'emailRetry'?: number;
    /**
     * 
     * @type {number}
     * @memberof NotificationEvent
     */
    'smsRetry'?: number;
    /**
     * 
     * @type {Timestamp}
     * @memberof NotificationEvent
     */
    'created'?: Timestamp;
    /**
     * 
     * @type {Timestamp}
     * @memberof NotificationEvent
     */
    'modified'?: Timestamp;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'creator'?: string;
    /**
     * 
     * @type {string}
     * @memberof NotificationEvent
     */
    'modifier'?: string;
}

export const Priority = {
    NotificationPriorityUnspecified: 'NOTIFICATION_PRIORITY_UNSPECIFIED',
    NotificationPriorityLow: 'NOTIFICATION_PRIORITY_LOW',
    NotificationPriorityMedium: 'NOTIFICATION_PRIORITY_MEDIUM',
    NotificationPriorityHigh: 'NOTIFICATION_PRIORITY_HIGH',
    NotificationPriorityImmediate: 'NOTIFICATION_PRIORITY_IMMEDIATE',
    Unrecognized: 'UNRECOGNIZED'
} as const;

export type Priority = typeof Priority[keyof typeof Priority];
export const Sensitivity = {
    SensitivityLevelUnspecified: 'SENSITIVITY_LEVEL_UNSPECIFIED',
    SensitivityLevelLow: 'SENSITIVITY_LEVEL_LOW',
    SensitivityLevelMedium: 'SENSITIVITY_LEVEL_MEDIUM',
    SensitivityLevelHigh: 'SENSITIVITY_LEVEL_HIGH',
    Unrecognized: 'UNRECOGNIZED'
} as const;

export type Sensitivity = typeof Sensitivity[keyof typeof Sensitivity];